requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
multidict==4.7.5
requests==2.23.0
six==1.14.0
slackclient==2.5.0
urllib3==1.25.8
websocket-client==0.54.0
yarl==1.4.2
